backend/requirements.txt
asgiref==3.9.1
azure-ai-agents==1.0.2
azure-ai-projects==1.0.0
azure-core==1.35.0
azure-cosmos==4.9.0
azure-identity==1.23.1
azure-storage-blob==12.26.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
Django==5.2.4
django-cors-headers==4.7.0
django-restframework==0.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
idna==3.10
isodate==0.7.2
msal==1.33.0
msal-extensions==1.3.1
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
urllib3==2.5.0
